feat(validation): update CarValidator to handle null price, adjust production year check, and switch price handling to BigDecimal
Enhanced `CarValidator` to include a null check for prices and adjust the production year validation to start from 1886. Updated price handling to use `BigDecimal` throughout codebase for better accuracy and consistency.

+++ src/main/java/de/ait/javalessonspro/repository/CarRepository.java
@@ -6,6 +6,7 @@ import de.ait.javalessonspro.model.Car;
 import lombok.NonNull;
 import org.springframework.data.jpa.repository.JpaRepository;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public interface CarRepository extends JpaRepository<Car, Long> {
@@ -19,7 +20,7 @@ public interface CarRepository extends JpaRepository<Car, Long> {
 
     List<Car> findByFuelType(FuelType fuelType);
 
-    List<Car> findByPriceBetween(double min, double max);
+    List<Car> findByPriceBetween(BigDecimal min, BigDecimal max);
 
     List<Car> findByHorsepowerBetween(int minHp, int maxHp);
